move duplicate to helper file
