+++ internal/app/services/blob_services_integration_test.go
@@ -12,8 +12,7 @@ import (
 	"crypto_vault_service/internal/infrastructure/settings"
 	"crypto_vault_service/internal/infrastructure/utils"
 	"crypto_vault_service/internal/persistence/repository"
-	"fmt"
-	"mime/multipart"
+	"crypto_vault_service/test/testutils"
 	"os"
 	"testing"
 
@@ -21,32 +20,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"gorm.io/gorm"
 )
-
-// Helper function to create test files
-func CreateTestFile(fileName string, content []byte) error {
-	err := os.WriteFile(fileName, content, 0600)
-	if err != nil {
-		return fmt.Errorf("failed to create test file: %w", err)
-	}
-	return nil
-}
-
-// Helper function to create a test file and form
-func CreateTestFileAndForm(t *testing.T, fileName string, fileContent []byte) (*multipart.Form, error) {
-	err := CreateTestFile(fileName, fileContent)
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		if err := os.Remove(fileName); err != nil {
-			t.Logf("failed to remove temporary file %s: %v", fileName, err)
-		}
-	})
-
-	form, err := utils.CreateForm(fileContent, fileName)
-	require.NoError(t, err)
-
-	return form, nil
-}
 
 type BlobServicesTest struct {
 	blobUploadService      blobs.BlobUploadService
@@ -116,7 +89,7 @@ func TestBlobUploadService_Upload_With_RSA_Encryption_And_Signing_Success(t *tes
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
 
-	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -148,7 +121,7 @@ func TestBlobUploadService_Upload_With_AES_Encryption_And_ECDSA_Signing_Success(
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
 
-	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -189,7 +162,7 @@ func TestBlobUploadService_Upload_Without_Encryption_And_Signing_Success(t *test
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
 
-	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -213,7 +186,7 @@ func TestBlobUploadService_Upload_Fail_InvalidEncryptionKey(t *testing.T) {
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
 
-	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -235,7 +208,7 @@ func TestBlobDownloadService_Download_Success(t *testing.T) {
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
 
-	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -279,7 +252,7 @@ func TestBlobMetadataService_List_Success(t *testing.T) {
 
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
-	err := CreateTestFile(testFileName, testFileContent)
+	err := testutils.CreateTestFile(testFileName, testFileContent)
 	require.NoError(t, err)
 	defer os.Remove(testFileName)
 
@@ -312,7 +285,7 @@ func TestBlobMetadataService_GetByID_Success(t *testing.T) {
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
 
-	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
@@ -339,7 +312,7 @@ func TestBlobMetadataService_DeleteByID_Success(t *testing.T) {
 	testFileContent := []byte("This is test file content")
 	testFileName := "testfile.txt"
 
-	form, err := CreateTestFileAndForm(t, testFileName, testFileContent)
+	form, err := testutils.CreateTestFileAndForm(t, testFileName, testFileContent)
 	require.NoError(t, err)
 
 	userId := uuid.New().String()
